Renamed test to testResources
