+++ .depcheckrc.yml
@@ -1,2 +1,3 @@
 ignores:
   - dotenv
+  - sort-package-json
