+++ resources/lint/profiles/mercury.camelcase-query-parameters.test.ts
@@ -12,7 +12,7 @@ import {
   conforms,
   breaksOnlyOneRule,
   renderSpecAsFile
-} from "../../../test/testUtils";
+} from "../../../testResources/testUtils";
 
 const PROFILE = "mercury";
