+++ resources/lint/profiles/mercury.mercury-profile.test.ts
@@ -12,7 +12,7 @@ import {
   breaksOnlyOneRule,
   renameKey,
   renderSpecAsFile
-} from "../../../test/testUtils";
+} from "../../../testResources/testUtils";
 
 const PROFILE = "mercury";
