+++ resources/lint/profiles/mercury.no-literal-question-marks-in-parameters.test.ts
@@ -12,7 +12,7 @@ import {
   conforms,
   breaksTheseRules,
   renderSpecAsFile
-} from "../../../test/testUtils";
+} from "../../../testResources/testUtils";
 
 const PROFILE = "mercury";
