+++ src/diff/apiDifferencer.test.ts
@@ -13,7 +13,7 @@ import tmp from "tmp";
 import { expect } from "chai";
 import { RuleCategory } from "./ruleSet";
 
-const basePath = path.join(__dirname, "../../test/diff");
+const basePath = path.join(__dirname, "../../testResources/diff");
 
 describe("Test RAML differencing", () => {
   it("can generate differences between RAML files", async () => {
