+++ src/download/exchangeDownloader.test.ts
@@ -14,7 +14,7 @@ import {
   search
 } from "./exchangeDownloader";
 import { RestApi } from "./exchangeTypes";
-import { searchAssetApiResultObject } from "../../test/download/resources/restApiResponseObjects";
+import { searchAssetApiResultObject } from "../../testResources/download/resources/restApiResponseObjects";
 
 import tmp from "tmp";
 import { Response } from "node-fetch";
@@ -25,13 +25,13 @@ import nock from "nock";
 import _ from "lodash";
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const assetSearchResults = require("../../test/download/resources/assetSearch.json");
+const assetSearchResults = require("../../testResources/download/resources/assetSearch.json");
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const getAssetWithVersion = require("../../test/download/resources/getAssetWithVersion");
+const getAssetWithVersion = require("../../testResources/download/resources/getAssetWithVersion");
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const getAssetWithoutVersion = require("../../test/download/resources/getAsset");
+const getAssetWithoutVersion = require("../../testResources/download/resources/getAsset");
 
 before(() => {
   chai.use(chaiAsPromised);
