+++ src/generate/api.test.ts
@@ -12,8 +12,14 @@ import chaiAsPromised from "chai-as-promised";
 import { Api } from "./api";
 import { model } from "amf-client-js";
 
-const validRamlFile = path.join(__dirname, "../../test/site.raml");
-const invalidRamlFile = path.join(__dirname, "../../test/search-invalid.raml");
+const validRamlFile = path.join(
+  __dirname,
+  "../../testResources/raml/site/site.raml"
+);
+const invalidRamlFile = path.join(
+  __dirname,
+  "../../testResources/raml/invalid/search-invalid.raml"
+);
 
 before(() => {
   chai.use(chaiAsPromised);
