+++ src/generate/apiCollection.test.ts
@@ -11,8 +11,14 @@ import chaiAsPromised from "chai-as-promised";
 
 import { ApiCollection } from "./apiCollection";
 
-const validRamlFile = path.join(__dirname, "../../test/site.raml");
-const invalidRamlFile = path.join(__dirname, "../../test/search-invalid.raml");
+const validRamlFile = path.join(
+  __dirname,
+  "../../testResources/raml/site/site.raml"
+);
+const invalidRamlFile = path.join(
+  __dirname,
+  "../../testResources/raml/invalid/search-invalid.raml"
+);
 
 before(() => {
   chai.use(chaiAsPromised);
