+++ src/generate/apiGroup.test.ts
@@ -13,8 +13,14 @@ import { model } from "amf-client-js";
 import { ApiGroup } from "./apiGroup";
 import { Api } from "./api";
 
-const validRamlFile = path.join(__dirname, "../../test/site.raml");
-const invalidRamlFile = path.join(__dirname, "../../test/search-invalid.raml");
+const validRamlFile = path.join(
+  __dirname,
+  "../../testResources/raml/site/site.raml"
+);
+const invalidRamlFile = path.join(
+  __dirname,
+  "../../testResources/raml/invalid/search-invalid.raml"
+);
 
 before(() => {
   chai.use(chaiAsPromised);
