+++ testResources/diff/left.raml
