+++ testResources/diff/right.raml
