+++ testResources/download/resources/no-category/raml-folder/exchange.json
