+++ testResources/generate/handlebarsAmfHelpersTestUtils.ts
