+++ testResources/raml/invalid/exchange.json
@@ -1,0 +1,12 @@
+{
+    "main": "search-invalid.raml",
+    "name": "Search Invalid",
+    "classifier": "raml",
+    "tags": [],
+    "groupId": "893f605e-10e2-423a-bdb4-f952f56eb6d8",
+    "backwardsCompatible": false,
+    "assetId": "search-invalid",
+    "version": "0.1.2",
+    "apiVersion": "v1",
+    "originalFormatVersion": "1.0"
+  }
