+++ testResources/raml/site/exchange.json
@@ -1,0 +1,12 @@
+{
+    "main": "site.raml",
+    "name": "Site",
+    "classifier": "raml",
+    "tags": [],
+    "groupId": "893f605e-10e2-423a-bdb4-f952f56eb6d8",
+    "backwardsCompatible": false,
+    "assetId": "sire",
+    "version": "0.1.2",
+    "apiVersion": "v1",
+    "originalFormatVersion": "1.0"
+  }
